Add component test
