--- app/tests/TestCase/Controller/Component/ArticlesComponentTest.php
@@ -1,45 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace App\Test\TestCase\Controller\Component;
-
-use App\Controller\Component\ArticlesComponent;
-use Cake\Controller\ComponentRegistry;
-use Cake\TestSuite\TestCase;
-
-/**
- * App\Controller\Component\ArticlesComponent Test Case
- */
-class ArticlesComponentTest extends TestCase
-{
-    /**
-     * Test subject
-     *
-     * @var \App\Controller\Component\ArticlesComponent
-     */
-    protected $Articles;
-
-    /**
-     * setUp method
-     *
-     * @return void
-     */
-    public function setUp(): void
-    {
-        parent::setUp();
-        $registry = new ComponentRegistry();
-        $this->Articles = new ArticlesComponent($registry);
-    }
-
-    /**
-     * tearDown method
-     *
-     * @return void
-     */
-    public function tearDown(): void
-    {
-        unset($this->Articles);
-
-        parent::tearDown();
-    }
-}
